feat: add extension methods for models
This adds extension methods for models such as `Create`, `Update`,
`Save` and `Delete`. Those methods are called upon the registered
repository. To use those methods, one needs to register the repositories
with the given `Add...` methods and call `UseModelExtensions` on the
`IServiceProvider`.

BREAKING CHANGE: Since it is best practice to use `Transient` lifetime
as much as possible, the default lifetime was changed from `Scoped` to
`Transient` in the registration methods.