feat: upgrade to ef core 3 (#4)
BREAKING CHANGE: This upgrades to netstandard2.1 and removes some
unused and ambiguous methods. First, it removes netcoreapp2.1 and netcoreapp2.2
as well as netstandard2.0. This is because ef core 3.0 depends on standard 2.1.

The methods for Query and QuerySingle were removed since they are ambiguous.
Please use AsQueryable().SingleAsync() or the general AsQueryable() instead.

The transactions are removed as well. Since the context does operate based on an instance
it could lead to errors. If you really need transactions, inject the context, perform the actions
and commit the transaction at the end.