feat: Improve database save / tracking logic (#6)
This MR refactors / improves the tracking logic of the EfRepositories.
In the past, all methods that used an enumerable of entites were saving after
each change made. This is refactored to use internal methods and saving
at the end of the database changes.

Also the determination if an object exists in the database was reworked to
take advantage of the Entities.Local cache. If objects are loaded
then tracked entities are checked first if they exist.